Edit the semantic error
The origin sentence is "But it does need the schema for that data, which includes information such as how many columns there are and which columns are numeric." 
In the sentence "it does need" means "필요하다" in Korean, but it is translated in apposite as "필요하지 않다(it does not need)". 
It should be updated.